+++ skoolkit/config.py
@@ -43,6 +43,7 @@ COMMANDS = {
         'CommentWidthMin': (10, ''),
         'DefbSize': (8, ''),
         'DefmSize': (66, ''),
+        'DefwSize': (1, ''),
         'InstructionWidth': (13, ''),
         'LineWidth': (79, 'line_width'),
         'ListRefs': (1, ''),